punten aantal hamer op barrel aangepast
de blauwe tonnen geven nu 600 ipv 300 zoals debruine tonnen geven

+++ Donkey_Kong/src/main/java/org/example/entities/obstakels/tonnen/BlauweTonnen.java
@@ -5,7 +5,7 @@ import com.github.hanyaeger.api.scenes.SceneBorder;
 
 public class BlauweTonnen extends Tonnen {
 
-    public BlauweTonnen(Coordinate2D initialLocation, int speed) {
+    public BlauweTonnen(Coordinate2D initialLocation, double speed) {
         super(initialLocation, speed);
         setMotion(speed, 270d);
     }
